Lesson 28) Activity: Project 1 "Star Break Coffee" Revenue/profit bar chart
The input data provided two numbers for each month: revenue and profit.  The instructions were not clear on which set to use, or both.  The lessons so far did not cover how to display multiple pieces of data from each row in a CSV.  The instructor's solution graphed revenue only.

This solution goes one step beyond, including both revenue and profit in a stacked bar graph.